Signal fixes, terminal fixes, sched fixes, ??? fixes, almost there boys

+++ kernel/lib/types.h
@@ -7,8 +7,8 @@
 typedef int64_t off_t;
 typedef int64_t ssize_t;
 
-typedef ssize_t pid_t;
-typedef ssize_t tid_t;
+typedef int pid_t;
+typedef int tid_t;
 
 typedef uint64_t dev_t;
 typedef uint64_t ino_t;
